major update: add Hibernate

+++ src/main/java/learn/epam/com/service/impl/TraineeServiceImpl.java
@@ -14,6 +14,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
@@ -32,6 +33,11 @@ public class TraineeServiceImpl implements TraineeService {
     private static final String AUTHENTICATION_FAIL = "Authentication failed";
     private static final String TRAINEE_ALREADY_ACTIVE = "Trainee already active";
     private static final String TRAINEE_ALREADY_INACTIVE = "Trainee already inactive";
+    private static final String USER_ID_REQUIRED = "Trainee.userId is required";
+    private static final String ADDRESS_REQUIRED = "Trainee.address is required";
+    private static final String DATE_OF_BIRTH_REQUIRED = "Trainee.dateOfBirth is required";
+    private static final String DATE_OF_BIRTH_IN_PAST_REQUIRED = "Trainee.dateOfBirth must be in the past";
+    private static final String ID_REQUIRED = "Trainee.id is required for update";
     private static final String NULL_EXCEPTION = "Argument is null ";
 
     private final TraineeDao traineeDao;
@@ -48,8 +54,11 @@ public class TraineeServiceImpl implements TraineeService {
     }
 
     @Override
+    @Transactional(rollbackFor = ServiceException.class)
     public void save(Trainee trainee) throws ServiceException {
         if (trainee != null) {
+            validateTraineeForCreate(trainee);
+
             userCredentialService.ensureUsernameExists(trainee.getUserId());
             userCredentialService.ensurePassword(trainee.getUserId());
 
@@ -62,13 +71,17 @@ public class TraineeServiceImpl implements TraineeService {
     }
 
     @Override
+    @Transactional
     public Optional<Trainee> findById(Long id) throws ServiceException {
         return traineeDao.getById(id);
     }
 
     @Override
+    @Transactional(rollbackFor = ServiceException.class)
     public void update(Trainee trainee) throws ServiceException {
         if (trainee != null) {
+            validateTraineeForUpdate(trainee);
+
             traineeDao.update(trainee);
 
             LOG.info(SUCCESS_UPDATE_TRAINEE);
@@ -78,6 +91,7 @@ public class TraineeServiceImpl implements TraineeService {
     }
 
     @Override
+    @Transactional(rollbackFor = ServiceException.class)
     public void delete(Trainee trainee) throws ServiceException {
         if (trainee != null) {
             traineeDao.delete(trainee);
@@ -89,11 +103,13 @@ public class TraineeServiceImpl implements TraineeService {
     }
 
     @Override
+    @Transactional
     public List<Trainee> findAllTrainee() {
         return traineeDao.getAll();
     }
 
     @Override
+    @Transactional
     public boolean checkCredentials(Long traineeId, String username, String password) throws ServiceException {
         if (traineeId != null && username != null && password != null) {
             Trainee trainee = findById(traineeId)
@@ -110,21 +126,28 @@ public class TraineeServiceImpl implements TraineeService {
     }
 
     @Override
+    @Transactional(rollbackFor = ServiceException.class)
     public Optional<Trainee> findTraineeByCredentials(String username, String password) throws ServiceException {
-        List<User> users = userDao.getAll();
-        return users.stream()
-                .filter(u -> username.equalsIgnoreCase(u.getUsername()) && password.equals(u.getPassword()))
-                .findFirst()
-                .flatMap(u ->
-                        traineeDao.getAll().stream()
-                                .filter(t -> t.getUserId().equals(u.getId()))
-                                .findFirst());
+        if (username != null & password != null) {
+            List<User> users = userDao.getAll();
+
+            return users.stream()
+                    .filter(u -> username.equalsIgnoreCase(u.getUsername()) && password.equals(u.getPassword()))
+                    .findFirst()
+                    .flatMap(u ->
+                            traineeDao.getAll().stream()
+                                    .filter(t -> t.getUserId().equals(u.getId()))
+                                    .findFirst());
+
+        } else {
+            throw new IllegalArgumentException(NULL_EXCEPTION);
+        }
     }
 
     @Override
+    @Transactional(rollbackFor = ServiceException.class)
     public Optional<Trainee> findTraineeByUsername(String username) {
         if (username != null) {
-
 
             return traineeDao.findTraineeByUsername(username);
         } else {
@@ -133,25 +156,29 @@ public class TraineeServiceImpl implements TraineeService {
     }
 
     @Override
+    @Transactional(rollbackFor = ServiceException.class)
     public void changePasswordForTrainee(String username, String oldPassword, String newPassword) throws ServiceException {
-        User user = userDao.getAll().stream()
-                .filter(user1 -> username.equalsIgnoreCase(user1.getUsername()))
-                .findFirst()
-                .orElseThrow(() -> new ServiceException(USER_NOT_FOUND));
+        if (username != null && oldPassword != null && newPassword != null) {
+            User user = userDao.getAll().stream()
+                    .filter(user1 -> username.equalsIgnoreCase(user1.getUsername()))
+                    .findFirst()
+                    .orElseThrow(() -> new ServiceException(USER_NOT_FOUND));
 
-        if (!user.getPassword().equals(oldPassword)) {
-            throw new ServiceException(INVALID_PASSWORD);
-        }
+            if (!user.getPassword().equals(oldPassword)) {
+                throw new ServiceException(INVALID_PASSWORD);
+            }
 
-        if (newPassword == null || newPassword.isBlank()) {
+            user.setPassword(newPassword);
+            userDao.update(user);
+        }else if (newPassword == null || newPassword.isBlank()) {
             throw new ServiceException(NEW_PASSWORD_REQUIRED);
+        } else {
+            throw new IllegalArgumentException(NULL_EXCEPTION);
         }
-
-        user.setPassword(newPassword);
-        userDao.update(user);
     }
 
     @Override
+    @Transactional(rollbackFor = ServiceException.class)
     public void updateTraineeProfile(String username, String password, Trainee updated) throws ServiceException {
         if (username != null && password != null && updated != null) {
             Trainee trainee = findTraineeByCredentials(username, password).orElseThrow(() -> new ServiceException(AUTHENTICATION_FAIL));
@@ -165,38 +192,52 @@ public class TraineeServiceImpl implements TraineeService {
     }
 
     @Override
+    @Transactional(rollbackFor = ServiceException.class)
     public void activateTrainee(String username) throws ServiceException {
-        Trainee trainee = findTraineeByUsername(username).orElseThrow(() -> new ServiceException(TRAINEE_NOT_FOUND));
-        if (trainee.isActive()) throw new ServiceException(TRAINEE_ALREADY_ACTIVE);
-        trainee.setActive(true);
-        traineeDao.update(trainee);
+        if (username != null) {
+            Trainee trainee = findTraineeByUsername(username).orElseThrow(() -> new ServiceException(TRAINEE_NOT_FOUND));
+            if (trainee.isActive()) throw new ServiceException(TRAINEE_ALREADY_ACTIVE);
+            trainee.setActive(true);
+            traineeDao.update(trainee);
+        } else {
+            throw new IllegalArgumentException(NULL_EXCEPTION);
+        }
     }
 
     @Override
+    @Transactional(rollbackFor = ServiceException.class)
     public void deactivateTrainee(String username) throws ServiceException {
-        Trainee trainee = findTraineeByUsername(username).orElseThrow(() -> new ServiceException(TRAINEE_NOT_FOUND));
-        if (!trainee.isActive()) throw new ServiceException(TRAINEE_ALREADY_INACTIVE);
-        trainee.setActive(false);
-        traineeDao.update(trainee);
+        if (username != null) {
+            Trainee trainee = findTraineeByUsername(username).orElseThrow(() -> new ServiceException(TRAINEE_NOT_FOUND));
+            if (!trainee.isActive()) throw new ServiceException(TRAINEE_ALREADY_INACTIVE);
+            trainee.setActive(false);
+            traineeDao.update(trainee);
+        } else {
+            throw new IllegalArgumentException(NULL_EXCEPTION);
+        }
     }
 
     @Override
-    @Transactional
+    @Transactional(rollbackFor = ServiceException.class)
     public void deleteTraineeByUsername(String username) throws ServiceException {
-        Trainee trainee = findTraineeByUsername(username)
-                .orElseThrow(() -> new ServiceException(AUTHENTICATION_FAIL));
+        if (username != null) {
+            Trainee trainee = findTraineeByUsername(username)
+                    .orElseThrow(() -> new ServiceException(AUTHENTICATION_FAIL));
 
-        Long traineeId = trainee.getId();
+            Long traineeId = trainee.getId();
 
-        trainingDao.getAll().stream()
-                .filter(training -> training.getTraineeId().equals(traineeId))
-                .forEach(training -> {
-                    trainingDao.delete(training);
-                });
+            trainingDao.getAll().stream()
+                    .filter(training -> training.getTraineeId().equals(traineeId))
+                    .forEach(training -> {
+                        trainingDao.delete(training);
+                    });
 
-        traineeDao.delete(trainee);
+            traineeDao.delete(trainee);
 
-        userDao.getById(trainee.getUserId()).ifPresent(userDao::delete);
+            userDao.getById(trainee.getUserId()).ifPresent(userDao::delete);
+        } else {
+            throw new IllegalArgumentException(NULL_EXCEPTION);
+        }
     }
 
     private User loadUser(Long userId) throws ServiceException {
@@ -206,5 +247,31 @@ public class TraineeServiceImpl implements TraineeService {
         } catch (ServiceException exception) {
             throw new ServiceException(FAIL_LOAD_USER, exception);
         }
+    }
+
+    private static void validateTraineeForCreate(Trainee trainee) throws ServiceException {
+        if (trainee != null) {
+            if (trainee.getUserId() == null) throw new ServiceException(USER_ID_REQUIRED);
+            if (isBlank(trainee.getAddress())) throw new ServiceException(ADDRESS_REQUIRED);
+            if (trainee.getDateOfBirth() == null) throw new ServiceException(DATE_OF_BIRTH_REQUIRED);
+            if (!trainee.getDateOfBirth().isBefore(LocalDate.now()))
+                throw new ServiceException(DATE_OF_BIRTH_IN_PAST_REQUIRED);
+        } else {
+            throw new IllegalArgumentException(NULL_EXCEPTION);
+        }
+    }
+
+    private static void validateTraineeForUpdate(Trainee trainee) throws ServiceException {
+        if (trainee != null) {
+            if (trainee.getId() == null) throw new ServiceException(ID_REQUIRED);
+            if (trainee.getUserId() == null) throw new ServiceException(USER_ID_REQUIRED);
+            validateTraineeForCreate(trainee);
+        } else {
+            throw new IllegalArgumentException(NULL_EXCEPTION);
+        }
+    }
+
+    private static boolean isBlank(String s) {
+        return s == null || s.isBlank();
     }
 }
